Fix finder query string with bad values
Avoid crash when query string for finder filters uses ids that are
not existing anymore (or when forging manually the query string).

It now check that the ids in the query string are accessible from the
current course too. Avoiding seeing some data from other courses.

If a filter has an invalid id, it remove it silently from the query
string, so the others ids (and filters) are not affected.