Fix MultiSelect promises never resolving
The default create, remove and change method in MultiSelect returned
promises that never resolve. This resulted in default behaviour (aka
when we don't override these methods) to never execute the callback
and thus it freezes the ReactSelect.

Used Promise.resolve() to fix the probleme.

"Promise.resolve()" is mostly a shorthand for
"new Promise((resolve) => resolve());"