Fix crash if presentation date is not valid
As the field is a string, it seems that on the legacy system, they are
some invalid date format.

To avoid the date picker component to crash in these cases, we
put a message to the user that the date is not valid and the
presentation date will be reseted at today.