Fix transcription max char on Chrome not always fulfilled
Sometimes the text wrap 1 char to early on Chrome. This behaviour is
very situational and inconsistent. It seems to appears when they are a
lot of spaces in the line. Sometime adding the last char on this line
will break, but othertime adding the same char at another position in
the line won't trigger the line break.

To fix this, we add a fraction of a char to the width of the textarea.