Use rawurlencode instead of urlencode
The main difference between the two is how spaces are encoded. urlencode
replace spaces with '+' while rawurlencode replace them with '%20'. In
our case, we need the raw version.

The url in the legacy version are encoded with raw, so we need to use
the same method to decode.

urlencode is mainly used for query params and rawurlencode should be
used for the part before query params.