Legacy migration (#105)
* Fix indentation

* Fix display glitch when state name is too long

* Remove unused permission

* Fix crash if presentation date is not valid

As the field is a string, it seems that on the legacy system, they are
some invalid date format.

To avoid the date picker component to crash in these cases, we
put a message to the user that the date is not valid and the
presentation date will be reseted at today.

* Migrate table users, courses, folders, states and cards

Cards data still in progress.

* Fix folder and card name with single quote

* Add heading plugin for editors

* Fix import button error when transcription is text

* Add text transcription importation for cards

* Fix transcription max char on Chrome not always fulfilled

Sometimes the text wrap 1 char to early on Chrome. This behaviour is
very situational and inconsistent. It seems to appears when they are a
lot of spaces in the line. Sometime adding the last char on this line
will break, but othertime adding the same char at another position in
the line won't trigger the line break.

To fix this, we add a fraction of a char to the width of the textarea.

* Fix multibyte string support in transcription validation

As char are multi byte encoded, strlen will return 2 with some char
like "ê".

Thanks PHP...

* Run Pint

* Add transcription importation

* Add font background color support in CK editor

Not added to the toolbar. It is just for retrocompatibility.
The toolbar has hidhlight tool that does mainly the same.

* Cut speaker to 3 characters if longer

* Add migration for enrollments (except cards)

* Sort tags by creation date

* Add migration for student_have_cards

* Add migration for tags

* Order admin courses list by name instead of created_at

* Add migration adding column legacy_id in courses, cards and folders

* Fill legacy id columns for cards, courses and folders

* Add an option to invalidate users email for testing purpose

* Add files and attachments migration

* Adapt tests not working anymore

* Fix typo

* Adapt FileStorageService for legacy file

* Add the possibility to add image by url for ck editor

* Increase filename length from 100 to 255 chars

* Use rawurlencode instead of urlencode

The main difference between the two is how spaces are encoded. urlencode
replace spaces with '+' while rawurlencode replace them with '%20'. In
our case, we need the raw version.

The url in the legacy version are encoded with raw, so we need to use
the same method to decode.

urlencode is mainly used for query params and rawurlencode should be
used for the part before query params.

* Rewrite url in box3 and box4 to match new url attachments

* Clean debug code

* Fix folder incorrect position

Because we update folders after creating them, triggering recalculation
of the position through folders observer.

* Export logs into a proper file