Do not take video_upload_state into account for files migation
In the legacy script migration, we taken the field video_upload_state
into consideration to know if the file had encoding problem.

But it did not works for every cases, this field is sometimes filled
with "Impossible d'encoder au format WEBM" even though the MP4 encoding
worked. As we use the MP4 encoding, this state is not relevent.

So now we completely ignore this field and just check the file on the
server if it exists or not to know if the video_url field is functional
or not.